test/fixtures/wpt/resource-timing/iframe-redirect-without-location.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test the sequence of events when reporting iframe timing.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/resource-loaders.js"></script>
<body>
<script>
    promise_test(async t => {
        const href = new URL('resources/redirect-without-location.py', location.href);
        await load.iframe(href);
        const entries = performance.getEntriesByType('resource').filter(({name}) => name.startsWith(href));
        assert_equals(entries.length, 1);
        assert_equals(entries[0].initiatorType, 'iframe');
    }, 'Iframes should report resource timing for redirect responses without a location');
</script>
</body>
